Try the `getMatrixFromRoot` approach to dealing with assembly contexts
